pyproject.toml
[tool.poetry]
name = "ms-ventas-crm"
version = "0.1.0"
description = "Microservicio FastAPI en Python 3.13 para Cloud Run"
authors = ["G17 <team@example.com>"]
package-mode = true

packages = [
  { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.110"
uvicorn = { extras = ["standard"], version = ">=0.29" }
pydantic = ">=2.7"
pydantic-settings = ">=2.2"
sqlalchemy = { extras = ["asyncio"], version = ">=2.0" }
asyncpg = ">=0.29"
redis = { extras = ["async"], version = ">=5.0" }
google-cloud-pubsub = ">=2.21"
google-cloud-bigquery = ">=3.25"
google-cloud-bigtable = ">=2.24"
google-cloud-storage = ">=2.18"
psycopg2-binary = "^2.9"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
pytest-cov = ">=5.0"
pytest-asyncio = ">=0.23"
httpx = ">=0.27"
ruff = ">=0.5"
requests = "^2.32.5"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--cov=src --cov-report=term-missing:skip-covered --cov-report=xml"
testpaths = ["tests"]